+++ sys/net/if_sl.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)if_sl.c	8.8 (Berkeley) 10/09/94
+ *	@(#)if_sl.c	8.9 (Berkeley) 01/09/95
  */
 
 /*
@@ -277,7 +277,7 @@ slclose(tp)
 int
 sltioctl(tp, cmd, data, flag)
 	struct tty *tp;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 {
@@ -766,7 +766,7 @@ newpack:
 int
 slioctl(ifp, cmd, data)
 	register struct ifnet *ifp;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 {
 	register struct ifaddr *ifa = (struct ifaddr *)data;
